refactor: replace & rename components from main tag
