+++ src/components/main/forms/Forms.tsx
@@ -1,0 +1,9 @@
+import './Forms.scss';
+
+import React from 'react';
+
+export default class Forms extends React.Component {
+  render() {
+    return <div>Forms</div>;
+  }
+}
