+++ src/components/main/home/Cards.scss
