+++ src/components/main/home/CardsInfo.tsx
