+++ src/components/main/home/Home.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
-import Main from './Main';
+import Main from './Home';
 
 describe('Search bar (Input)', () => {
   it('Main page should contain search bar', () => {
